requirements.txt
Flask==2.3.2
openpyxl==3.1.2
Pillow==10.0.0